Misc cleanup and restructuring of handlers after reorg
