+++ public/app/features/alerting/unified/components/receivers/form/GrafanaReceiverForm.test.tsx
@@ -36,7 +36,6 @@ const ui = {
 
 describe('GrafanaReceiverForm', () => {
   beforeEach(() => {
-    server.resetHandlers();
     clearPluginSettingsCache();
   });
