+++ public/app/features/alerting/unified/components/receivers/grafanaAppReceivers/onCall/useOnCallIntegration.test.ts
@@ -15,7 +15,6 @@ const server = setupMswServer();
 
 describe('useOnCallIntegration', () => {
   afterEach(() => {
-    server.resetHandlers();
     clearPluginSettingsCache();
   });
