+++ public/app/features/alerting/unified/testSetup/plugins.ts
@@ -3,7 +3,7 @@ import { RequestHandler } from 'msw';
 import { PluginMeta, PluginType } from '@grafana/data';
 import { config } from '@grafana/runtime';
 
-import { pluginsHandler } from '../mocks/plugins';
+import { getPluginsHandler } from '../mocks/server/handlers/plugins';
 
 export function setupPlugins(plugins: PluginMeta[]): { apiHandlers: RequestHandler[] } {
   plugins.forEach((plugin) => {
@@ -17,7 +17,7 @@ export function setupPlugins(plugins: PluginMeta[]): { apiHandlers: RequestHandl
   });
 
   return {
-    apiHandlers: [pluginsHandler(plugins)],
+    apiHandlers: [getPluginsHandler(plugins)],
   };
 }
 
@@ -90,5 +90,28 @@ export const plugins: PluginMeta[] = [
     },
     module: 'public/plugins/grafana-asserts-app/module.js',
     baseUrl: 'public/plugins/grafana-asserts-app',
+  },
+  {
+    id: 'grafana-oncall-app',
+    name: 'OnCall',
+    type: PluginType.app,
+    enabled: true,
+    info: {
+      author: {
+        name: 'Grafana Labs',
+        url: '',
+      },
+      description: 'OnCall',
+      links: [],
+      logos: {
+        small: '',
+        large: '',
+      },
+      screenshots: [],
+      version: 'local-dev',
+      updated: '2024-04-09',
+    },
+    module: 'public/plugins/grafana-oncall-app/module.js',
+    baseUrl: 'public/plugins/grafana-oncall-app',
   },
 ];
